pyproject.toml
[tool.poetry]
name = "mar-2023-pyton-drf444"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-filter = "^23.3"
pillow = "^10.1.0"
django-cors-headers = "^4.3.0"
drf-yasg = "^1.21.7"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"
psycopg2 = "^2.9.9"
drf-api-logger = "^1.1.15"
python-dotenv = "^1.0.0"
django-extensions = "^3.2.3"
pandas = "^2.1.3"
free-proxy = "^1.1.1"
grequests = "^0.7.0"
ipython = "^8.18.1"
scrapy-djangoitem = "^1.1.1"
scrapy = "^2.11.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
g4f = "0.1.9.3"
graphene-django = "^3.1.5"
django-graphql-jwt = "^0.4.0"
pytest-playwright = "^0.4.3"


[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
